+++ targets/EDA_plot_targets.R
@@ -40,8 +40,7 @@ EDA_plot_targets <- list(
           model == "base",
           n_history_delay == 60,
           n_history_uncertainty == 60,
-          borrow_delay == FALSE,
-          borrow_uncertainty == FALSE
+          borrow == FALSE
         ),
       final_summed_data = final_eval_data_covid_7d,
       nowcast_dates_to_plot = c("2021-12-01", "2022-02-01", "2022-04-01"),